Remove max_page_size
This parameter affects existing API calls.
While I do not think it will cause any issues, I am removing
it for now because I want the changes to be isolated to the
new API as much as possible. I'd be happy to reintroduce this
later as-is, or before release if it targets this API specifically.