Amend asset rewrite implementation  (#613)
* Amend subdomain asset rewrite implementation

Why
---

For MVP2, we implemented an aspect of subdomain updates by rewriting
assets inside a hub instance. The implementation created copies of files
before performing the subdomain rewrite. The previous versions of the
assets would have been auto-deleted eventually, but this is an issue if
a user updates their subdomain frequently, since it would accumulate
older assets over and over again. This will also cause metrics to be
overblown until the auto-delete occurs.

It’s probably best if we fix this now before MVP2 users run into it.

What
----

* Remove old scene assets after rewrite
* Remove old project assets after rewrite
* Test that owned files are rewritten
* Move database variables out to the host OS for test config

Co-authored-by: Bryan Enders <2033828+bryanenders@users.noreply.github.com>