Check for the global name if the per-server nickname isn't present
What: If no per-server nickname has been set, this checks to see if Discord's global name is set for the user before falling back to the username.

Why: This makes the name used in Hubs match what Discord displays in the Discord server instead of predominantly using their (usually less aesthetically pleasing) username.

Notes:
It is unlikely that the username will ever be used because the global name should always be set.