Fix user permissions bug for Discord bound rooms
What: Fixes a bug with getting the permissions of users for Discord bound rooms.  The permissions for the roles a user has weren't getting combined into a single bitfield correctly and instead it seemed that the permissions for just one of the roles was being picked (which may or may not give them all the powers they are supposed to have).  This explicitly specifies the order of operations so that everything gets combined correctly.

Why: Users weren't always given the powers they should have been based on their Discord roles, which could prevent moderators and admins from being able to kick users and edit room settings.

Notes:
The bug was introduced when the Discord API version was updated and the handling for Discord sometimes sending strings and sometimes sending integers was added.

Commit that introduced the bug: 886394a503864435a8abedcc304ebc0d01d59ed8